refactor: load the AI prompt form on demand
